Moves pagination helpers back out of application_helper and into pagination_helper.
